// src/domain/types/register-state.ts
export interface RegisterState {
  username: string;
  name: string;
  lastName: string;
  location: string;
  alternativeLocation: string;
  gender: string;
  otherGender: string;
  age: string;
  email: string;
  password: string;
  confirmPassword: string;
  error: string;
  backPressedOnce: boolean;
}